python/requirements.txt
Flask
flask-sqlalchemy
gevent
psycopg2-binary
